+++ src/Components/Login/Login.css
@@ -129,88 +129,3 @@
     text-decoration: underline;
     color: rgb(24, 228, 41);
 }
-
-.logRegContent .regContent{
-    position: a;
-}
-
-.regContent{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding-left: 55px;
-}
-
-.regContent .regImg img{
-    width: 90%;
-    margin-left: 10%;
-}
-
-.regForm{
-    width: 500px;
-    color: #fff;
-    border-radius: 30px;
-    background: transparent;
-    border: 3px solid rgba(255, 255, 255, 0.2);
-    backdrop-filter: blur(30px);
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
-    padding: 30px 40px;
-    margin-right: 10%;
-    text-align: center;
-}
-
-.regForm h1{
-    font-size: 40px;
-    text-align: center;
-    text-transform: uppercase;
-}
-
-.regForm .inputBox{
-    position: relative;
-    width: 100%;
-    height: 50px;
-    margin: 30px 0;
-}
-
-.regForm .agree{
-    font-size: 15px;
-}
-
-.agree label input{
-    accent-color: #fff;
-    margin-right: 8px;
-}
-
-.agree label{
-    font-size: 18px;
-}
-
-.regForm button{
-    width: 40%;
-    height: 50px;
-    background: #fff;
-    border: none;
-    outline: none;
-    border-radius: 40px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    cursor: pointer;
-    font-size: 22px;
-    color: #121316;
-    font-weight: 700;
-    margin-top: 5%;
-}
-
-.regForm button:hover{
-    background-color: rgb(24, 228, 41);
-}
-
-.regForm .loginLink p{
-    font-size: 18px;
-    text-align: center;
-    margin: 20px 0 15px;
-}
-
-.loginLink p a:hover{
-    text-decoration: underline;
-    color: rgb(24, 228, 41);
-}
